refactor: Refatorando controllers
Refatoração do response dos controllers, para melhor legibilidade e manutenabilidade

+++ src/controllers/goal.controller.ts
@@ -1,6 +1,12 @@
 import type { Request, Response } from "express";
-import { responseMessages } from "../constants/messages.constants.js";
 import type GoalService from "../services/goal.service.js";
+import errorResponseWith from "../utils/errorResponseWith.js";
+import successResponseWith from "../utils/successResponseWith.js";
+import isMissingFields from "../utils/isMissingFields.js";
+import {
+  ARBITRARY_FIELDS_MESSAGE,
+  MISSING_FIELDS_MESSAGE,
+} from "../constants/messages.constants.js";
 
 class GoalController {
   private goalService: GoalService;
@@ -13,29 +19,42 @@ class GoalController {
     try {
       const allGoalsData = await this.goalService.getAllGoalsData();
 
-      return res.status(200).json(allGoalsData);
+      return res
+        .status(200)
+        .json(
+          successResponseWith(allGoalsData, "Dados encontrados com sucesso."),
+        );
     } catch (err) {
-      return res.status(500).json({
-        message: responseMessages.catchErrorMessage,
-        error: (err as Error).message,
-      });
+      const error = err as Error;
+      return res.status(500).json(errorResponseWith(error.message, 500));
     }
   }
 
   async createNewGoal(req: Request, res: Response): Promise<Response> {
     try {
       const newGoalInfos = req.body;
+      const fields = Object.keys(newGoalInfos);
+
+      if (isMissingFields(newGoalInfos)) {
+        return res
+          .status(422)
+          .json(
+            errorResponseWith(
+              ARBITRARY_FIELDS_MESSAGE(fields),
+              422,
+              MISSING_FIELDS_MESSAGE,
+            ),
+          );
+      }
 
       const newGoal = await this.goalService.createNewGoal(newGoalInfos);
 
       return res
         .status(201)
-        .json({ message: "Meta criada com sucess.", goal: newGoal });
+        .json(successResponseWith(newGoal, "Meta criada com sucesso.", 201));
     } catch (err) {
-      return res.status(500).json({
-        message: responseMessages.catchErrorMessage,
-        error: (err as Error).message,
-      });
+      const error = err as Error;
+      return res.status(500).json(errorResponseWith(error.message, 500));
     }
   }
 
@@ -43,14 +62,31 @@ class GoalController {
     try {
       const { uuid } = req.params;
 
+      if (!uuid) {
+        return res
+          .status(422)
+          .json(
+            errorResponseWith(
+              ARBITRARY_FIELDS_MESSAGE(["uuid"]),
+              422,
+              MISSING_FIELDS_MESSAGE,
+            ),
+          );
+      }
+
       await this.goalService.deleteGoal(uuid as string);
 
-      return res.status(200).json({ message: "Meta removida com sucesso." });
+      return res
+        .status(200)
+        .json(
+          successResponseWith(
+            "Remoção concluída",
+            "Meta removida com sucesso.",
+          ),
+        );
     } catch (err) {
-      return res.status(500).json({
-        message: responseMessages.catchErrorMessage,
-        error: (err as Error).message,
-      });
+      const error = err as Error;
+      return res.status(500).json(errorResponseWith(error.message, 500));
     }
   }
 
@@ -58,6 +94,31 @@ class GoalController {
     try {
       const updateGoalValues = req.body;
       const { uuid } = req.params;
+      const fields = Object.keys(updateGoalValues);
+
+      if (!uuid) {
+        return res
+          .status(422)
+          .json(
+            errorResponseWith(
+              ARBITRARY_FIELDS_MESSAGE(["uuid"]),
+              422,
+              MISSING_FIELDS_MESSAGE,
+            ),
+          );
+      }
+
+      if (isMissingFields(updateGoalValues)) {
+        return res
+          .status(422)
+          .json(
+            errorResponseWith(
+              ARBITRARY_FIELDS_MESSAGE(fields),
+              422,
+              MISSING_FIELDS_MESSAGE,
+            ),
+          );
+      }
 
       const updatedGoal = await this.goalService.updateGoalData(
         updateGoalValues,
@@ -66,12 +127,10 @@ class GoalController {
 
       return res
         .status(200)
-        .json({ message: "Dados de meta atualizado.", update: updatedGoal });
+        .json(successResponseWith(updatedGoal, "Dados de meta atualizados."));
     } catch (err) {
-      return res.status(500).json({
-        message: responseMessages.catchErrorMessage,
-        error: (err as Error).message,
-      });
+      const error = err as Error;
+      return res.status(500).json(errorResponseWith(error.message, 500));
     }
   }
 }
